refactor: move WalletConnect.js to WalletConnect folder
