+++ app/core/WalletConnect/WalletConnect.js
@@ -1,26 +1,29 @@
 import RNWalletConnect from '@walletconnect/client';
 import { v1 as random } from 'uuid';
-import Engine from './Engine';
-import Logger from '../util/Logger';
+import Engine from '../Engine';
+import Logger from '../../util/Logger';
 // eslint-disable-next-line import/no-nodejs-modules
 import { EventEmitter } from 'events';
 import AsyncStorage from '@react-native-async-storage/async-storage';
-import { CLIENT_OPTIONS, WALLET_CONNECT_ORIGIN } from '../util/walletconnect';
-import { WALLETCONNECT_SESSIONS } from '../constants/storage';
+import {
+  CLIENT_OPTIONS,
+  WALLET_CONNECT_ORIGIN,
+} from '../../util/walletconnect';
+import { WALLETCONNECT_SESSIONS } from '../../constants/storage';
 import { WalletDevice } from '@metamask/transaction-controller';
-import BackgroundBridge from './BackgroundBridge/BackgroundBridge';
+import BackgroundBridge from '../BackgroundBridge/BackgroundBridge';
 import getRpcMethodMiddleware, {
   checkActiveAccountAndChainId,
   ApprovalTypes,
-} from './RPCMethods/RPCMethodMiddleware';
+} from '../RPCMethods/RPCMethodMiddleware';
 import { Linking } from 'react-native';
 import Minimizer from 'react-native-minimizer';
-import AppConstants from './AppConstants';
-import { strings } from '../../locales/i18n';
-import NotificationManager from './NotificationManager';
-import { msBetweenDates, msToHours } from '../util/date';
+import AppConstants from '../AppConstants';
+import { strings } from '../../../locales/i18n';
+import NotificationManager from '../NotificationManager';
+import { msBetweenDates, msToHours } from '../../util/date';
 import URL from 'url-parse';
-import parseWalletConnectUri from './WalletConnect/wc-utils';
+import parseWalletConnectUri from './wc-utils';
 
 const hub = new EventEmitter();
 let connectors = [];
