+++ app/core/WalletConnect/WalletConnect.test.ts
